+++ tket/src/Circuit/Multiplexor.cpp
@@ -1008,9 +1008,8 @@ void add_multi_rz(
   TKET_ASSERT(!all_decomps.empty());
   unsigned reference_size = 0;
   for (unsigned i = 0; i < all_decomps.size(); i++) {
-    if (!all_decomps[i].empty() && !reference_size) {
-      reference_size = all_decomps[i].size();
-    }
+    if (all_decomps[i].empty()) continue;
+    if (!reference_size) reference_size = all_decomps[i].size();
     TKET_ASSERT(reference_size == all_decomps[i].size());
   }